Define it before you use it!
Good coding practice generally is that if fun2 calls fun1, fun1 should be defined prior to fun2 in the file. In some languages it's the difference between compilation and non-compilation; while R is not one of those the most readable code is the code that fits the expectations of the most people, so I've rearranged so that lsatGET and similar are defined prior, not after, the functions using them.

(Similarly, included httr:: explicitly to make clear where functions are coming from to anyone reading the code. Plus, CRAN likes it.)